Introduce a namespace abstraction to simplify endpoint registration.
Earlier versions of this library put a lot of responsiblity into the Operation enum
to define naming conventions. Since the naming depends heavily on the context of the
resources being operated on, the Operation functions had some fairly complex argumentation.

In the new approach, these functions move to a Namespace class, which has concrete
instance variables to make the naming evaluation simpler. As much as possible, convention
creation functions now accept Namespace instances, though not yet everywhere because of
backward compatibility concerns.