Support null valued fields in swagger responses.
Swagger apparently does not support null values; JSON Schema does, but nullability
needs to be expressed as a `oneOf` clause over both the `null` type and the actual
type of the object; Swagger appears to have eliminated this complexity with the
expectation that null values are just removed from the data type.

This change does a few things:

 - It supports pruning null values from microcosm JSON response; this behavior
   is conditional on the `X-Response-Skip-Null` header being supplied.

 - It injects this header as an optional parameter of every microcosm request.

 - It marks `allow_none` (marshmallow) fields as non-required Swagger fields.

From a Swagger client (e.g. Yelp's bravado), we can now handle null fields via:

    response = client.foo.bar(
        args,
        _request_options=dict(
            headers={
               "X-Response-Skip-Null": True,
            },
        ),
    )