Correctly handle schema with no required fields in swagger definition.
Previously, we were inserting a `required` attribute into every definition,
even if it had no required fields. This created problems with schema validation,
which expected the `required` value to be a non-empty list. Our workaround was
introducing the `EmptySchema` base with a placeholder required field.

It turns out that we can just omit the `required` field entirely if there are
no required fields. This makes swagger usage (and validation) much simpler.